% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMonSearchInterface.R
\name{SMRMonCreateSearchInteractiveInterface}
\alias{SMRMonCreateSearchInteractiveInterface}
\title{Creation of a SMRMon search interface}
\usage{
SMRMonCreateSearchInteractiveInterface(
  smrObj,
  itemData,
  itemDataColNames = NULL,
  itemDataIDColName = NULL,
  itemListIDsSplitPattern = "\\\\W",
  dashboardTheme = NULL
)
}
\arguments{
\item{smrObj}{A SMRMon object.}

\item{itemData}{A data frame with rows corresponding to items (to be recommended.)}

\item{itemDataColNames}{Which column names of \code{itemData} should be used in
the search and recommendation results.}

\item{itemDataIDColName}{Which column of \code{itemData} is with item ID's;
those ID's are also row names of \code{smrObj$M}.}

\item{itemListIDsSplitPattern}{A split pattern for the separator of the items list
and ratings list.}

\item{dashboardTheme}{A string specifying the dashbaard theme.
See \code{\link{dashboardthemes::shinyDashboardThemes}}.
Ignored if the package \code{dashboardthemes} is not installed.}
}
\value{
Shiny app
}
\description{
Creation of a SMRMon search interface
}
\details{
The default value of \code{itemListIDsSplitPattern} is '\\W', but "," should be
used if the row ID's have white spaces in them.
}
\seealso{
Other SMRMon interface functions: 
\code{\link{SMRMonCreateInteractiveInterface}()},
\code{\link{SMRMonMakeSearchServerFunction}()},
\code{\link{SMRMonMakeSearchUI}()},
\code{\link{SMRMonMakeUI}()}
}
\concept{SMRMon interface functions}
