% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMonToLSAMon.R
\name{SMRMonToLSAMon}
\alias{SMRMonToLSAMon}
\title{SMRMon to LSAMon}
\usage{
SMRMonToLSAMon(
  smr,
  splitPattern = "[[:space:]]|[[:punct:]]",
  stemWordsQ = FALSE,
  stemRules = NULL,
  stopWords = NULL,
  numberOfTopics = 60,
  minNumberOfDocumentsPerTerm = 0,
  sep = ":"
)
}
\arguments{
\item{smr}{A SMRMon object.}

\item{splitPattern}{Pattern to use for splitting the documents into words.
If NULL \code{"[[:space:]]|[[:punct:]]"} is used.}

\item{stemWordsQ}{Should the words be stemmed or not?
If \code{stemWordsQ = TRUE} and \code{stemRules = NULL}
then \code{\link{SnowballC::wordStem}} is used.}

\item{stemRules}{A character vector with named elements or NULL.
(A vector of "stemming rules".)
If \code{stemWordsQ = FALSE} then \code{stemRules} is ignored.}

\item{stopWords}{A character vector with stop words to be removed.}

\item{numberOfTopics}{Number of topics to be extracted.
If less or equal to 0 no topic extraction is made.}

\item{minNumberOfDocumentsPerTerm}{Minimal number of documents for the terms
to be considered in the topics.}

\item{sep}{Separator of tag type and tag in the columns \code{smr$M}.}
}
\value{
A LSAMon object
}
\description{
Make an LSAMon object from an SMRMon object.
}
\details{
Takes as arguments the arguments of \code{\link{LSAMonMakeDocumentTermMatrix}}
and \code{\link{LSAMonExtractTopics}}.
}
