% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesSearchInterface.R
\name{TSCorrSMRExtendRecommendations}
\alias{TSCorrSMRExtendRecommendations}
\title{Extend time series recommendations}
\usage{
TSCorrSMRExtendRecommendations(tsSMR, recs, roundDigits = 6)
}
\arguments{
\item{tsSMR}{A time series recommender.}

\item{recs}{A data frame with columns \"Score\" and \"Item\".}

\item{roundDigits}{Number of decimal places for \code{\link{round}}.
(Used for making the \code{ggplot} panel names.)}
}
\value{
A data frame
}
\description{
Extends the time series recommendations data frame \code{recs} with
time axis and value axis data from the time series recommender object \code{tsSMR}.
The obtained time series data frame is in long form.
}
\seealso{
Other Time series search interface functions: 
\code{\link{TSCorrSMRCreateSearchInterface}()},
\code{\link{TSCorrSMRMakeServerFunction}()},
\code{\link{TSCorrSMRMakeUI}()},
\code{\link{TSCorrSMRPlotRecommendations}()}
}
\concept{Time series search interface functions}
