% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesSearchInterface.R
\name{TSCorrSMRPlotRecommendations}
\alias{TSCorrSMRPlotRecommendations}
\title{Plot time series recommendations}
\usage{
TSCorrSMRPlotRecommendations(
  tsSMR,
  recs,
  roundDigits = 6,
  ncol = NULL,
  scales = "fixed"
)
}
\arguments{
\item{tsSMR}{A time series recommender.}

\item{recs}{A data frame with columns \"Score\" and \"Item\".}

\item{roundDigits}{Number of decimal places for \code{\link{round}}.
(Used for making the \code{ggplot} panel names.)}

\item{ncol}{Number of columns argument for \code{\link{ggplot2::facet_wrap}}.}

\item{scales}{Scales argument for \code{\link{ggplot2::facet_wrap}}.}
}
\value{
A \code{ggplot} object
}
\description{
Makes time series plots from the
time series recommendations data frame \code{recs} using
time axis and value axis data from the time series recommender object \code{tsSMR}.
}
\details{
Extends the recommendations with \code{\link{TSCorrSMRExtendRecommendations}}.
}
\seealso{
Other Time series search interface functions: 
\code{\link{TSCorrSMRCreateSearchInterface}()},
\code{\link{TSCorrSMRExtendRecommendations}()},
\code{\link{TSCorrSMRMakeServerFunction}()},
\code{\link{TSCorrSMRMakeUI}()}
}
\concept{Time series search interface functions}
