% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesSearchInterface.R
\name{TSCorrSMRMakeUI}
\alias{TSCorrSMRMakeUI}
\title{Time series search interface UI}
\usage{
TSCorrSMRMakeUI(
  tsSMR,
  tsSearchVectors,
  initNNs = 12,
  initNCols = 2,
  dashboardTitle = "Time series search engine",
  plotOutputHeight = "1000px",
  noteText = NULL,
  theme = "cerulean"
)
}
\arguments{
\item{tsSMR}{A time series recommender.}

\item{tsSearchVectors}{A list of time series search vectors.}

\item{initNNs}{Initial number nearest neighbors.}

\item{initNCols}{Initial number of columns.}

\item{dashboardTitle}{Dashboard title.}

\item{plotOutputHeight}{Plot output height for \code{\link{plotOutput}}.}

\item{noteText}{Note text to display.}

\item{theme}{Shiny-theme to apply.}
}
\value{
Shiny UI object.
}
\description{
Creates the Shiny UI function for a time series search interface.
}
\seealso{
Other Time series search interface functions: 
\code{\link{TSCorrSMRCreateSearchInterface}()},
\code{\link{TSCorrSMRMakeServerFunction}()}
}
\concept{Time series search interface functions}
